# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "User API",
    "version": "1.0.0",
    "description": "API documentation for user authentication, management, and access control"
  },
  "paths": {
...
    "/content-blocks": {
      "get": {
        "summary": "Get all content blocks",
        "tags": ["Content Blocks"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "List of all content blocks" }
        }
      }
    },
    "/content/block": {
      "post": {
        "summary": "Create a new content block",
        "tags": ["Content Blocks"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "parentId": { "type": "integer" },
                  "order": { "type": "integer" },
                  "code": { "type": "string" },
                  "content": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "language": {
                          "type": "string",
                          "enum": ["ru", "ua", "en", "tr", "de", "es"],
                          "default": "ua"
                        },
                        "text": { "type": "string" },
                        "mediaUrl": { "type": "string" },
                        "mediaType": {
                          "type": "string",
                          "enum": ["photo", "video", "document", "gif", "none"],
                          "default": "none"
                        }
                      },
                      "required": ["language", "text"]
                    }
                  }
                },
                "required": ["parentId", "content"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Content block created successfully" },
          "404": { "description": "Error creating content block" }
        }
      }
    },
    "/content/block/translation/{blockId}/{lang}": {
      "post": {
        "summary": "Create or update translation for a content block",
        "tags": ["Content Blocks"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "blockId",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" },
            "description": "ID of the content block"
          },
          {
            "name": "lang",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "enum": ["ru", "ua", "en", "tr", "de", "es"]
            },
            "description": "Language code for the translation"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "text": {
                    "type": "string",
                    "example": "Translated text here"
                  },
                  "mediaType": {
                    "type": "string",
                    "enum": ["none", "photo", "video", "document", "gif"],
                    "example": "none"
                  },
                  "mediaFile": {
                    "type": "string",
                    "format": "binary",
                    "description": "Optional media file for the translation"
                  }
                },
                "required": ["text", "mediaType"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Translation created or updated successfully"
          },
          "403": { "description": "Block not found or access denied" },
          "500": { "description": "Server error" }
        }
      }
    },
    "/content/block/{blockId}": {
      "get": {
        "summary": "Get a content block by ID",
        "tags": ["Content Blocks"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "blockId",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": { "description": "Content block data" },
          "404": { "description": "Content block not found" }
        }
      },
      "put": {
        "summary": "Update an existing content block",
        "tags": ["Content Blocks"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "blockId",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "parentId": { "type": "integer" },
                  "order": { "type": "integer" },
                  "code": { "type": "string" },
                  "content": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "language": {
                          "type": "string",
                          "enum": ["ru", "ua", "en", "tr", "de", "es"],
                          "default": "ua"
                        },
                        "text": { "type": "string" },
                        "mediaUrl": { "type": "string" },
                        "mediaType": {
                          "type": "string",
                          "enum": ["photo", "video", "document", "gif", "none"],
                          "default": "none"
                        }
                      },
                      "required": ["language", "text"]
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Content block updated successfully" },
          "404": { "description": "Error updating content block" }
        }
      },
      "delete": {
        "summary": "Delete a content block",
        "tags": ["Content Blocks"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "blockId",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": { "description": "Content block deleted successfully" },
          "404": { "description": "Error deleting content block" }
        }
      }
    },
    "/content/button": {
      "post": {
        "summary": "Create a new content button",
        "tags": ["Content Buttons"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "blockId": { "type": "integer" },
                  "order": { "type": "integer" },
                  "type": {
                    "type": "string",
                    "enum": ["url", "callback", "web_app", "none"],
                    "default": "none"
                  },
                  "keyboardType": {
                    "type": "string",
                    "enum": ["inline", "reply"]
                  },
                  "isFullWidth": { "type": "boolean" },
                  "rowOrder": { "type": "integer" },
                  "nextBlockId": { "type": "integer" },
                  "label": { "type": "string" },
                  "url": { "type": "string" },
                  "translations": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "language": {
                          "type": "string",
                          "enum": ["ru", "ua", "en", "tr", "de", "es"],
                          "default": "ua"
                        },
                        "label": { "type": "string" }
                      },
                      "required": ["language", "label"]
                    }
                  }
                },
                "required": ["label", "blockId"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Content button created successfully" },
          "404": { "description": "Error creating content button" }
        }
      }
    },
    "/content/button/{buttonId}": {
      "get": {
        "summary": "Get a content button by ID",
        "tags": ["Content Buttons"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "buttonId",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": { "description": "Content button data" },
          "404": { "description": "Content button not found" }
        }
      },
      "put": {
        "summary": "Update an existing content button",
        "tags": ["Content Buttons"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "buttonId",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "blockId": { "type": "integer" },
                  "order": { "type": "integer" },
                  "type": {
                    "type": "string",
                    "enum": ["url", "callback", "web_app", "none"],
                    "default": "none"
                  },
                  "keyboardType": {
                    "type": "string",
                    "enum": ["inline", "reply"]
                  },
                  "isFullWidth": { "type": "boolean" },
                  "rowOrder": { "type": "integer" },
                  "nextBlockId": { "type": "integer" },
                  "label": { "type": "string" },
                  "url": { "type": "string" },
                  "callback": { "type": "string" },
                  "translations": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "language": {
                          "type": "string",
                          "enum": ["ru", "ua", "en", "tr", "de", "es"],
                          "default": "ua"
                        },
                        "label": { "type": "string" }
                      },
                      "required": ["language", "label"]
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Content button updated successfully" },
          "404": { "description": "Error updating content button" }
        }
      },
      "delete": {
        "summary": "Delete a content button",
        "tags": ["Content Buttons"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "buttonId",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": { "description": "Content button deleted successfully" },
          "404": { "description": "Error deleting content button" }
        }
      }
    }
  },
  "components": {
...
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}